requirements.txt
aiohttp==3.6.2
allensdk==2.2.0
argon2-cffi==20.1.0
argschema==1.17.5
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
black==20.8b1
bleach==3.2.1
bson==0.5.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
entrypoints==0.3
future==0.18.2
Glymur==0.8.19
POT==0.7.0
pymongo==3.11.1
h5netcdf==0.8.1
h5py==2.10.0
hdmf==2.2.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.0.0rc6
matplotlib==3.3.2
mistune==0.8.4
mock==4.0.2
-e git+https://github.com/neuroailab/NormativeMEC.git#egg=mec
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.2.0
networkx==2.5
notebook==6.1.4
numexpr==2.7.1
numpy==1.18.5
opencv-python==4.1.1.26
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pynrrd==0.4.2
pynwb==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-build==0.11.1
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
shapely==1.7.1
SimpleITK==1.2.4
simplejson==3.17.2
six==1.15.0
sklearn==0.0
statsmodels==0.9.0
tables==3.5.1
tensorflow-gpu==2.0.0rc2
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xarray==0.15.1
yarl==1.6.0
zipp==3.2.0
